requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1
django-cors-headers==4.3.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
idna==3.10
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
